backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "A FastAPI backend for the project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.12",
    "networkx==3.4.1",
    "pillow>=11.2.1",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.51.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.6",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu124", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu124", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
